+++ src/LightOps.CQRS/Api/Commands/ICommand.cs
@@ -1,5 +1,8 @@
 ﻿namespace LightOps.CQRS.Api.Commands
 {
+    /// <summary>
+    /// Marker interface for commands
+    /// </summary>
     public interface ICommand
     {